+++ json-typedef/java-jackson/AssetControlFieldEnhancementConditionMeter.java
@@ -5,16 +5,10 @@ package Datasworn;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
-/**
- * Some assets provide a special condition meter of their own. The most common
- * example is the health meters on companion assets. Asset condition meters
- * may also include their own controls, such as the checkboxes that Starforged
- * companion assets use to indicate they are "out of action".
- */
 @JsonSerialize
 public class AssetControlFieldEnhancementConditionMeter extends AssetControlFieldEnhancement {
     @JsonProperty("max")
-    private Short max;
+    private Byte max;
 
     public AssetControlFieldEnhancementConditionMeter() {
     }
@@ -23,7 +17,7 @@ public class AssetControlFieldEnhancementConditionMeter extends AssetControlFiel
      * Getter for max.<p>
      * The maximum value of this meter.
      */
-    public Short getMax() {
+    public Byte getMax() {
         return max;
     }
 
@@ -31,7 +25,7 @@ public class AssetControlFieldEnhancementConditionMeter extends AssetControlFiel
      * Setter for max.<p>
      * The maximum value of this meter.
      */
-    public void setMax(Short max) {
+    public void setMax(Byte max) {
         this.max = max;
     }
 }
